make BridgeBuilder a subclass of ThingBuilder and adapt BaseBridgeHandler.editThing()
