+++ bundles/core/org.eclipse.smarthome.core.thing/src/main/java/org/eclipse/smarthome/core/thing/binding/BaseBridgeHandler.java
@@ -12,6 +12,7 @@ import java.util.List;
 import org.eclipse.smarthome.core.thing.Bridge;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingUID;
+import org.eclipse.smarthome.core.thing.binding.builder.BridgeBuilder;
 
 /**
  * {@link BaseBridgeHandler} adds some convenience methods for bridges to the {@link BaseThingHandler}.
@@ -53,4 +54,18 @@ public abstract class BaseBridgeHandler extends BaseThingHandler {
     public Bridge getThing() {
         return (Bridge) super.getThing();
     }
+
+    /**
+     * Creates a bridge builder, which allows to modify the bridge. The method
+     * {@link BaseThingHandler#updateThing(Thing)} must be called to persist the changes.
+     *
+     * @return {@link BridgeBuilder} which builds an exact copy of the bridge (not null)
+     */
+    @Override
+    protected BridgeBuilder editThing() {
+        return (BridgeBuilder) BridgeBuilder.create(this.thing.getThingTypeUID(), this.thing.getUID())
+                .withBridge(this.thing.getBridgeUID()).withChannels(this.thing.getChannels())
+                .withConfiguration(this.thing.getConfiguration());
+    }
+
 }
