+++ bundles/core/org.eclipse.smarthome.core.thing/src/main/java/org/eclipse/smarthome/core/thing/binding/builder/BridgeBuilder.java
@@ -9,12 +9,20 @@ package org.eclipse.smarthome.core.thing.binding.builder;
 
 import java.util.ArrayList;
 
+import org.eclipse.smarthome.core.thing.Bridge;
 import org.eclipse.smarthome.core.thing.Channel;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.ThingUID;
 import org.eclipse.smarthome.core.thing.internal.BridgeImpl;
 
-public class BridgeBuilder extends GenericBridgeBuilder<BridgeBuilder> {
+/**
+ * This class allows the easy construction of a {@link Bridge} instance using the builder pattern.
+ *
+ * @author Dennis Nobel - Initial contribution and API
+ * @author Kai Kreuzer - Refactoring to make BridgeBuilder a subclass of ThingBuilder
+ *
+ */
+public class BridgeBuilder extends ThingBuilder {
 
     private BridgeBuilder(BridgeImpl thing) {
         super(thing);
@@ -37,4 +45,8 @@ public class BridgeBuilder extends GenericBridgeBuilder<BridgeBuilder> {
         return new BridgeBuilder(bridge);
     }
 
+    @Override
+    public Bridge build() {
+        return (Bridge) super.build();
+    }
 }
