+++ bundles/core/org.eclipse.smarthome.core.thing/src/main/java/org/eclipse/smarthome/core/thing/binding/builder/ThingBuilder.java
@@ -7,14 +7,31 @@
  */
 package org.eclipse.smarthome.core.thing.binding.builder;
 
+import java.util.List;
+import java.util.Map;
+
+import org.eclipse.smarthome.config.core.Configuration;
+import org.eclipse.smarthome.core.thing.Channel;
+import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.ThingUID;
 import org.eclipse.smarthome.core.thing.internal.ThingImpl;
 
-public class ThingBuilder extends GenericThingBuilder<ThingBuilder> {
+import com.google.common.collect.Lists;
 
-    private ThingBuilder(ThingImpl thing) {
-        super(thing);
+/**
+ * This class allows the easy construction of a {@link Thing} instance using the builder pattern.
+ *
+ * @author Dennis Nobel - Initial contribution and API
+ * @author Kai Kreuzer - Refactoring to make BridgeBuilder a subclass
+ *
+ */
+public class ThingBuilder {
+
+    private ThingImpl thing;
+
+    protected ThingBuilder(ThingImpl thing) {
+        this.thing = thing;
     }
 
     public static ThingBuilder create(ThingTypeUID thingTypeUID, String thingId) {
@@ -31,6 +48,57 @@ public class ThingBuilder extends GenericThingBuilder<ThingBuilder> {
     public static ThingBuilder create(ThingTypeUID thingTypeUID, ThingUID thingUID) {
         ThingImpl thing = new ThingImpl(thingTypeUID, thingUID);
         return new ThingBuilder(thing);
+    }
+
+    public ThingBuilder withLabel(String label) {
+        this.thing.setLabel(label);
+        return this;
+    }
+
+    public ThingBuilder withChannel(Channel channel) {
+        List<Channel> channels = this.thing.getChannelsMutable();
+        channels.add(channel);
+        return this;
+    }
+
+    public ThingBuilder withChannels(Channel... channels) {
+        this.thing.setChannels(Lists.newArrayList(channels));
+        return this;
+    }
+
+    public ThingBuilder withChannels(List<Channel> channels) {
+        this.thing.setChannels(Lists.newArrayList(channels));
+        return this;
+    }
+
+    public ThingBuilder withConfiguration(Configuration thingConfiguration) {
+        this.thing.setConfiguration(thingConfiguration);
+        return this;
+    }
+
+    public ThingBuilder withBridge(ThingUID bridgeUID) {
+        if (bridgeUID != null) {
+            this.thing.setBridgeUID(bridgeUID);
+        }
+        return this;
+    }
+
+    public ThingBuilder withProperties(Map<String, String> properties) {
+        if (properties != null) {
+            for (String key : properties.keySet()) {
+                this.thing.setProperty(key, properties.get(key));
+            }
+        }
+        return this;
+    }
+
+    public ThingBuilder withLocation(String location) {
+        this.thing.setLocation(location);
+        return this;
+    }
+
+    public Thing build() {
+        return this.thing;
     }
 
 }
