+++ bundles/core/org.eclipse.smarthome.core.thing/src/main/java/org/eclipse/smarthome/core/thing/util/ThingHelper.java
@@ -18,7 +18,6 @@ import org.eclipse.smarthome.core.thing.Channel;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingUID;
 import org.eclipse.smarthome.core.thing.binding.builder.BridgeBuilder;
-import org.eclipse.smarthome.core.thing.binding.builder.GenericThingBuilder;
 import org.eclipse.smarthome.core.thing.binding.builder.ThingBuilder;
 import org.eclipse.smarthome.core.thing.dto.ChannelDTO;
 import org.eclipse.smarthome.core.thing.dto.ChannelDTOMapper;
@@ -106,7 +105,7 @@ public class ThingHelper {
      */
     public static Thing merge(Thing thing, ThingDTO updatedContents) {
 
-        GenericThingBuilder<?> builder;
+        ThingBuilder builder;
 
         if (thing instanceof Bridge) {
             builder = BridgeBuilder.create(thing.getThingTypeUID(), thing.getUID());
